Refactor CLI frontend parsing and execution
