+++ crates/cli/src/frontend/command_builder/mod.rs
@@ -2,7 +2,7 @@ mod sections;
 
 pub(super) use clap::{Arg, ArgAction, Command as ClapCommand, builder::OsStringValueParser};
 
-pub(super) fn clap_command(program_name: &'static str) -> ClapCommand {
+pub(crate) fn clap_command(program_name: &'static str) -> ClapCommand {
     let command = sections::section_01(program_name);
     let command = sections::section_02(command);
     sections::section_03(command)
